# fix: improve parsing content-range header