--- a/lib/core/util.js
+++ b/lib/core/util.js
@@ -487,19 +487,41 @@ function isValidHTTPToken (characters) {
   return true
 }
 
-// Parsed accordingly to RFC 9110
-// https://www.rfc-editor.org/rfc/rfc9110#field.content-range
-function parseRangeHeader (range) {
-  if (range == null || range === '') return { start: 0, end: null, size: null }
+const emptyRange = Object.freeze({ start: 0, end: null, size: null })
+const rangeHeaderRE = /^bytes (?:(?:(?:(\d+)-(\d+)))\/(?:\*|(\d+))$|\*\/(\d+))$/
 
-  const m = range ? range.match(/^bytes (\d+)-(\d+)\/(\d+)?$/) : null
-  return m
-    ? {
-        start: parseInt(m[1]),
-        end: m[2] ? parseInt(m[2]) : null,
-        size: m[3] ? parseInt(m[3]) : null
-      }
-    : null
+/**
+ * @typedef {Object} SatisfiedContentRangeHeader
+ * @property {number} start
+ * @property {number} end
+ * @property {number | null} size
+ */
+
+/**
+ * @typedef {Object} UnsatisfiedContentRangeHeader
+ * @property {0} start
+ * @property {null} end
+ * @property {number} size
+ */
+
+/**
+ * Parsed accordingly to RFC 9110
+ *
+ * @param {null|string} range
+ * @returns {null|SatisfiedContentRangeHeader|UnsatisfiedContentRangeHeader}
+ * @see https://www.rfc-editor.org/rfc/rfc9110#field.content-range
+ */
+function parseContentRangeHeader (range) {
+  if (range == null || range === '') {
+    return emptyRange
+  }
+
+  const m = range.match(rangeHeaderRE)
+  return m && ((m[4] && { start: 0, end: null, size: +m[4] }) || {
+    start: +m[1],
+    end: (m[2] && +m[2]) ?? null,
+    size: (m[3] && +m[3]) ?? null
+  })
 }
 
 const kEnumerableProperty = Object.create(null)
@@ -538,7 +560,7 @@ module.exports = {
   addAbortListener,
   isValidHTTPToken,
   isTokenCharCode,
-  parseRangeHeader,
+  parseContentRangeHeader,
   nodeMajor,
   nodeMinor,
   nodeHasAutoSelectFamily: nodeMajor > 18 || (nodeMajor === 18 && nodeMinor >= 13),