--- a/lib/handler/RetryHandler.js
+++ b/lib/handler/RetryHandler.js
@@ -2,7 +2,7 @@ const assert = require('node:assert')
 
 const { kRetryHandlerDefaultRetry } = require('../core/symbols')
 const { RequestRetryError } = require('../core/errors')
-const { isDisturbed, parseHeaders, parseRangeHeader } = require('../core/util')
+const { isDisturbed, parseHeaders, parseContentRangeHeader } = require('../core/util')
 
 function calculateRetryAfterHeader (retryAfter) {
   const current = Date.now()
@@ -58,7 +58,13 @@ class RetryHandler {
     }
 
     this.retryCount = 0
+    /**
+     * @type {number}
+     */
     this.start = 0
+    /**
+     * @type {number | null}
+     */
     this.end = null
     this.etag = null
     this.resume = null
@@ -198,9 +204,9 @@ class RetryHandler {
         return true
       }
 
-      const contentRange = parseRangeHeader(headers['content-range'])
+      const contentRange = parseContentRangeHeader(headers['content-range'])
       // If no content range
-      if (!contentRange) {
+      if (contentRange === null) {
         this.abort(
           new RequestRetryError('Content-Range mismatch', statusCode, {
             headers,
@@ -224,7 +230,7 @@ class RetryHandler {
       const { start, size, end = size } = contentRange
 
       assert(this.start === start, 'content-range mismatch')
-      assert(this.end == null || this.end === end, 'content-range mismatch')
+      assert(this.end === null || this.end === end, 'content-range mismatch')
 
       this.resume = resume
       return true
@@ -233,9 +239,9 @@ class RetryHandler {
     if (this.end == null) {
       if (statusCode === 206) {
         // First time we receive 206
-        const range = parseRangeHeader(headers['content-range'])
+        const contentRange = parseContentRangeHeader(headers['content-range'])
 
-        if (range == null) {
+        if (contentRange === null) {
           return this.handler.onHeaders(
             statusCode,
             rawHeaders,
@@ -244,15 +250,14 @@ class RetryHandler {
           )
         }
 
-        const { start, size, end = size } = range
+        const { start, size, end = size } = contentRange
 
         assert(
-          start != null && Number.isFinite(start) && this.start !== start,
+          this.start !== start,
           'content-range mismatch'
         )
-        assert(Number.isFinite(start))
         assert(
-          end != null && Number.isFinite(end) && this.end !== end,
+          end !== null && this.end !== end,
           'invalid content-length'
         )
 
@@ -268,7 +273,7 @@ class RetryHandler {
 
       assert(Number.isFinite(this.start))
       assert(
-        this.end == null || Number.isFinite(this.end),
+        this.end === null || Number.isFinite(this.end),
         'invalid content-length'
       )