@Uzlopak 2024-02-19
I will first implement some unit tests, as some of the expectations to the code seem to be wrong.