@metcoder95 2024-02-20
Sadly I don't see much room for improving the current `range` parsing. The only improvement I see is to replace it with the current implementation with the `fetch` one and adjust accordingly, so we have a single implementation shared across and we can optimize that single implementation and have benefits at different ends.

Regarding `unsatisfied ranges`, I believe that `416` status code should result in an Aborted request and give back control to the implementation; I am not so sure what else we can do in those scenarios beyond stopping the request and not retrying more.