@metcoder95 2024-02-20
Sure, that SGTM 👍